+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/component/table/FilteredTableTest.java
@@ -46,7 +46,7 @@ public class FilteredTableTest {
 
 		FilteredTableModel<List<String>, Integer> tableModel =
 						FilteredTableModel.<List<String>, Integer>builder(() -> singletonList(column), List::get)
-										.itemSupplier(() -> asList(
+										.items(() -> asList(
 														singletonList("darri"),
 														singletonList("dac"),
 														singletonList("dansinn"),